requirements.txt
pyspark==3.4.1
flask==2.3.3
flask-cors==4.0.0
flask-caching==2.1.0
pandas==2.0.3
numpy==1.25.2
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
requests==2.31.0
hdfs3==0.3.1
urllib3==2.0.4
pathlib==1.0.1
python-dateutil==2.8.2